integrations: minimize and migrate to BATS
Delete tests that seem to have minimal utility, and move the rest to
the end-to-end BATS suite as appropriate. All but one test in the
integrations suite have been removed as a result. (The remaining test
seemed easier to keep as Go code due to the helpful behavior of
goroutines.)

The copy-paste "updates status PENDING to RUNNING then to COMPLETE if
successful" tests and friends were deleted with prejudice and not
replaced. In general, they added "coverage" without actually validating
useful behavior, and the mock gymnastics performed to get there were way
too expensive.

By far the most complex migration was for init-cluster. Hopefully some
other improvements we hope to make in the future should make testing
much easier for that particular step.

The suite-specific code has also been minimized. To help future test
writers, we assume all executables under test are already on the PATH,
and to help developers who might forget about that, it bails on all
tests if the three gpupgrade binaries aren't already there. It also
warns if those binaries appear to be in an installed location instead of
the local development directory.

The README has been updated with some cautionary examples and opinions
on the place of unit, integration, and end-to-end tests.

+++ test/config.bats
@@ -3,18 +3,26 @@
 load helpers
 
 setup() {
+    skip_if_no_gpdb
+
     STATE_DIR=`mktemp -d`
     export GPUPGRADE_HOME="${STATE_DIR}/gpupgrade"
 
     kill_agents
     kill_hub
-    gpupgrade initialize --old-bindir "${GPHOME}/bin" --new-bindir "${GPHOME}/bin" --old-port ${PGPORT}
+    gpupgrade initialize \
+        --old-bindir "${GPHOME}/bin" \
+        --new-bindir "${GPHOME}/bin" \
+        --old-port ${PGPORT} 3>&-
 }
 
 teardown() {
-    kill_agents
-    kill_hub
-    rm -r "${STATE_DIR}"
+    # XXX Beware, BATS_TEST_SKIPPED is not a documented export.
+    if [ -z "${BATS_TEST_SKIPPED}" ]; then
+        kill_hub
+        kill_agents
+        rm -r "${STATE_DIR}"
+    fi
 }
 
 @test "configuration can be read after it is written" {
@@ -31,18 +39,15 @@ teardown() {
     [ "$output" = "/my/old/bin/dir" ]
 }
 
-# TODO: we need to rewrite this test after we work through the design
-#   of how `gpupgrade recover` will work
-#
 @test "configuration persists after hub is killed and restarted" {
-   gpupgrade config set --new-bindir /my/bin/dir
-   kill_hub
+    gpupgrade config set --new-bindir /my/bin/dir
+    kill_hub
 
-   gpupgrade_hub --daemonize
+    gpupgrade_hub --daemonize
 
-   run gpupgrade config show --new-bindir
-   [ "$status" -eq 0 ]
-   [ "$output" = "/my/bin/dir" ]
+    run gpupgrade config show --new-bindir
+    [ "$status" -eq 0 ]
+    [ "$output" = "/my/bin/dir" ]
 }
 
 @test "configuration can be dumped as a whole" {
